+++ src/main/scala/au/com/agl/arc/util/ConfigUtils.scala
@@ -2293,6 +2293,45 @@ object ConfigUtils {
     }
   }  
 
+  def readTextLoad(idx: Int, graph: Graph, name: StringConfigValue, params: Map[String, String])(implicit spark: SparkSession, logger: au.com.agl.arc.util.log.logger.Logger, c: Config): (Either[List[StageError], PipelineStage], Graph) = {
+    import ConfigReader._
+
+    val expectedKeys = "type" :: "name" :: "description" :: "environments" :: "inputView" :: "outputURI" :: "authentication" :: "numPartitions" :: "partitionBy" :: "saveMode" :: "params" :: "singleFile" :: "prefix" :: "separator" :: "suffix" :: Nil
+    val invalidKeys = checkValidKeys(c)(expectedKeys)  
+
+    val description = getOptionalValue[String]("description")
+
+    val inputView = getValue[String]("inputView") |> graph.vertexExists("inputView") _
+    val outputURI = getValue[String]("outputURI") |> validateURI("outputURI") _
+    val numPartitions = getOptionalValue[Int]("numPartitions")
+    val authentication = readAuthentication("authentication")  
+    val saveMode = getValue[String]("saveMode", default = Some("Overwrite"), validValues = "Append" :: "ErrorIfExists" :: "Ignore" :: "Overwrite" :: Nil) |> parseSaveMode("saveMode") _
+
+    val singleFile = getValue[Boolean]("singleFile", default = Some(false))
+    val prefix = getValue[String]("prefix", default = Some(""))
+    val separator = getValue[String]("separator", default = Some(""))
+    val suffix = getValue[String]("suffix", default = Some(""))
+
+    (name, description, inputView, outputURI, numPartitions, authentication, saveMode, invalidKeys, singleFile, prefix, separator, suffix) match {
+      case (Right(n), Right(d), Right(iv), Right(out), Right(np), Right(auth), Right(sm), Right(_), Right(sf), Right(pre), Right(sep), Right(suf)) => 
+        val load = TextLoad(n, d, iv, out, np, auth, sm, params, sf, pre, sep, suf)
+
+        val ov = s"$idx:${load.getType}"
+        var outputGraph = graph
+        // add the vertices
+        outputGraph = outputGraph.addVertex(Vertex(idx, ov))
+        // add the edges
+        outputGraph = outputGraph.addEdge(iv, ov)
+
+        (Right(load), outputGraph)
+      case _ =>
+        val allErrors: Errors = List(name, description, inputView, outputURI, numPartitions, authentication, saveMode, invalidKeys, singleFile, prefix, separator, suffix).collect{ case Left(errs) => errs }.flatten
+        val stageName = stringOrDefault(name, "unnamed stage")
+        val err = StageError(idx, stageName, c.origin.lineNumber, allErrors)
+        (Left(err :: Nil), graph)
+    }
+  } 
+
   def readXMLLoad(idx: Int, graph: Graph, name: StringConfigValue, params: Map[String, String])(implicit spark: SparkSession, logger: au.com.agl.arc.util.log.logger.Logger, c: Config): (Either[List[StageError], PipelineStage], Graph) = {
     import ConfigReader._
 
@@ -2622,6 +2661,7 @@ object ConfigUtils {
           case Right("KafkaLoad") => readKafkaLoad(idx, graph, name, params)
           case Right("ORCLoad") => readORCLoad(idx, graph, name, params)
           case Right("ParquetLoad") => readParquetLoad(idx, graph, name, params)
+          case Right("TextLoad") => readTextLoad(idx, graph, name, params)
           case Right("XMLLoad") => readXMLLoad(idx, graph, name, params)
 
           case Right("HTTPExecute") => readHTTPExecute(idx, graph, name, params)