+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added `DatabricksSQLDWLoad` stage for bulk loading Azure SQLDW when executing in the [Databricks Runtime](https://databricks.com/product/databricks-runtime) environment.
 - added `ElasticsearchExtract` and `ElasticsearchLoad` stages for connecting to [Elasticsearch](https://www.elastic.co/products/elasticsearch) clusters.
 - added additional checks for table dependencies when validating the job config.
+- added `TextLoad` which supports both `singleFile` and standard partitioned output formats.
 
 # 1.8.0